refactor: remove legacy limits checker and rename new one
