+++ test/unit/domain/accounts/limits-checker.spec.ts
@@ -1,73 +1,107 @@
 import { getAccountLimits } from "@config"
-import { LimitsChecker } from "@domain/accounts"
+import { AccountLimitsChecker } from "@domain/accounts"
 import { LimitsExceededError } from "@domain/errors"
-import { toCents } from "@domain/fiat"
+import { PriceRatio } from "@domain/payments"
+import {
+  AmountCalculator,
+  paymentAmountFromNumber,
+  WalletCurrency,
+  ZERO_CENTS,
+} from "@domain/shared"
+
+let usdPaymentAmount: UsdPaymentAmount
+let limitsChecker: AccountLimitsChecker
+let walletVolumeIntraledger: TxBaseVolumeAmount<WalletCurrency>
+let walletVolumeWithdrawal: TxBaseVolumeAmount<WalletCurrency>
+let priceRatio: PriceRatio
+
+const calc = AmountCalculator()
+
+const ONE_CENT = { amount: 1n, currency: WalletCurrency.Usd } as UsdPaymentAmount
+
+beforeAll(async () => {
+  const priceRatioResult = PriceRatio({
+    usd: ONE_CENT,
+    btc: { amount: 50n, currency: WalletCurrency.Btc },
+  })
+  if (priceRatioResult instanceof Error) throw priceRatioResult
+  priceRatio = priceRatioResult
+
+  const level: AccountLevel = 1
+  const accountLimits = getAccountLimits({ level })
+
+  usdPaymentAmount = {
+    amount: 10_000n,
+    currency: WalletCurrency.Usd,
+  }
+
+  limitsChecker = AccountLimitsChecker({
+    accountLimits,
+    priceRatio,
+  })
+
+  const intraLedgerOutgoingBaseAmount = paymentAmountFromNumber({
+    amount: accountLimits.intraLedgerLimit - Number(usdPaymentAmount.amount),
+    currency: WalletCurrency.Usd,
+  })
+  if (intraLedgerOutgoingBaseAmount instanceof Error) throw intraLedgerOutgoingBaseAmount
+  walletVolumeIntraledger = {
+    outgoingBaseAmount: intraLedgerOutgoingBaseAmount,
+    incomingBaseAmount: ZERO_CENTS,
+  }
+
+  const withdrawalOutgoingBaseAmount = paymentAmountFromNumber({
+    amount: accountLimits.withdrawalLimit - Number(usdPaymentAmount.amount),
+    currency: WalletCurrency.Usd,
+  })
+  if (withdrawalOutgoingBaseAmount instanceof Error) throw withdrawalOutgoingBaseAmount
+  walletVolumeWithdrawal = {
+    outgoingBaseAmount: withdrawalOutgoingBaseAmount,
+    incomingBaseAmount: ZERO_CENTS,
+  }
+})
 
 describe("LimitsChecker", () => {
-  let paymentAmount: UsdCents
-  let limitsChecker: LimitsChecker
-  let walletVolumeIntraledger: TxBaseVolume
-  let walletVolumeWithdrawal: TxBaseVolume
-  beforeAll(() => {
-    const level: AccountLevel = 1
-    const accountLimits = getAccountLimits({ level })
-
-    paymentAmount = toCents(10_000)
-    limitsChecker = LimitsChecker({
-      accountLimits,
-    })
-
-    walletVolumeIntraledger = {
-      outgoingBaseAmount: toCents(accountLimits.intraLedgerLimit - paymentAmount),
-      incomingBaseAmount: toCents(0),
-    }
-
-    walletVolumeWithdrawal = {
-      outgoingBaseAmount: toCents(accountLimits.withdrawalLimit - paymentAmount),
-      incomingBaseAmount: toCents(0),
-    }
-  })
-
-  it("passes for exact limit amount", () => {
-    const intraledgerLimitCheck = limitsChecker.checkIntraledger({
-      amount: paymentAmount,
-      walletVolume: walletVolumeIntraledger,
+  it("passes for exact limit amount", async () => {
+    const intraledgerLimitCheck = await limitsChecker.checkIntraledger({
+      amount: usdPaymentAmount,
+      walletVolumes: [walletVolumeIntraledger],
     })
     expect(intraledgerLimitCheck).not.toBeInstanceOf(Error)
 
-    const withdrawalLimitCheck = limitsChecker.checkWithdrawal({
-      amount: paymentAmount,
-      walletVolume: walletVolumeWithdrawal,
+    const withdrawalLimitCheck = await limitsChecker.checkWithdrawal({
+      amount: usdPaymentAmount,
+      walletVolumes: [walletVolumeWithdrawal],
     })
     expect(withdrawalLimitCheck).not.toBeInstanceOf(Error)
   })
 
-  it("passes for amount below limit", () => {
-    const intraledgerLimitCheck = limitsChecker.checkIntraledger({
-      amount: toCents(paymentAmount - 1),
-      walletVolume: walletVolumeIntraledger,
+  it("passes for amount below limit", async () => {
+    const intraledgerLimitCheck = await limitsChecker.checkIntraledger({
+      amount: calc.sub(usdPaymentAmount, ONE_CENT),
+      walletVolumes: [walletVolumeIntraledger],
     })
     expect(intraledgerLimitCheck).not.toBeInstanceOf(Error)
 
-    const withdrawalLimitCheck = limitsChecker.checkWithdrawal({
-      amount: toCents(paymentAmount - 1),
-      walletVolume: walletVolumeWithdrawal,
+    const withdrawalLimitCheck = await limitsChecker.checkWithdrawal({
+      amount: calc.sub(usdPaymentAmount, ONE_CENT),
+      walletVolumes: [walletVolumeWithdrawal],
     })
     expect(withdrawalLimitCheck).not.toBeInstanceOf(Error)
   })
 
-  it("returns an error for exceeded intraledger amount", () => {
-    const intraledgerLimitCheck = limitsChecker.checkIntraledger({
-      amount: toCents(paymentAmount + 1),
-      walletVolume: walletVolumeIntraledger,
+  it("returns an error for exceeded intraledger amount", async () => {
+    const intraledgerLimitCheck = await limitsChecker.checkIntraledger({
+      amount: calc.add(usdPaymentAmount, ONE_CENT),
+      walletVolumes: [walletVolumeIntraledger],
     })
     expect(intraledgerLimitCheck).toBeInstanceOf(LimitsExceededError)
   })
 
-  it("returns an error for exceeded withdrawal amount", () => {
-    const withdrawalLimitCheck = limitsChecker.checkWithdrawal({
-      amount: toCents(paymentAmount + 1),
-      walletVolume: walletVolumeWithdrawal,
+  it("returns an error for exceeded withdrawal amount", async () => {
+    const withdrawalLimitCheck = await limitsChecker.checkWithdrawal({
+      amount: calc.add(usdPaymentAmount, ONE_CENT),
+      walletVolumes: [walletVolumeWithdrawal],
     })
     expect(withdrawalLimitCheck).toBeInstanceOf(LimitsExceededError)
   })
